+++ CHANGELOG.md
@@ -8,3 +8,5 @@ This project uses [Semantic Versioning 2.0.0](http://semver.org/).
 Initial import from the `whois` library.
 
 - NEW: whois.dk-hostmaster.dk parser now recognizes throttled responses (whois/GH-382). [Thanks @troelskn]
+
+- FIXED: whois.audns.net.au removed the registrar ID field (GH-20, GH-21). Thanks [@afoster]